Refactor - Move SelectMany to a different file
